+++ src/parser/enrichers/monster/ConjuredFey/_module.mjs
@@ -1,0 +1,1 @@
+export { default as PsychicAttack } from "./PsychicAttack.mjs";