+++ src/parser/enrichers/monster/_module.mjs
@@ -4,3 +4,4 @@ export * as PurpleWorm from "./PurpleWorm/_module.mjs";
 export * as Dragon from "./Dragon/_module.mjs";
 export * as ConjuredAnimals from "./ConjuredAnimals/_module.mjs";
 export * as ConjuredElemental from "./ConjuredElemental/_module.mjs";
+export * as ConjuredFey from "./ConjuredFey/_module.mjs";